+++ 0x0F-function_pointers/function_pointers.h
@@ -2,7 +2,7 @@
 #define FUNCTION_POINTERS
 int _putchar(char c);
 void print_name(char *name, void (*f)(char *));
-void array_iterator(int *array, unsigned int size, void (*action)(int));
+void array_iterator(int *array, size_t size, void (*action)(int));
 int int_index(int *array, int size, int (*cmp)(int));
 int (*get_op_func(char *s))(int, int);
 #endif
